feat: Add ability to manage Trivy binary from plugin
Resolves #41

When Trivy can't be found on the path then give an option to download
and install Trivy using the plugin. This version will be managed by the
plugin and store the file in the plugins directory.

If the path of Trivy is the plugin folder and Trivy is installed, an
Update Trivy button will be provided.

This PR has an additional small fix to allow selecting text in the
findings helper which Resolves #43